--- src/output/io.h
@@ -1,20 +1,0 @@
-#ifndef INCLUDE_IO_H
-#define INCLUDE_IO_H
-
-/** outb:
-*  Sends the given data to the given I/O port. Defined in io.s
-*
-*  @param port The I/O port to send the data to
-*  @param data The data to send to the I/O port
-*/
-extern "C" void outb(unsigned short port, unsigned char data);
-
-/** inb:
-*  Read a byte from an I/O port.
-*
-*  @param  port The address of the I/O port
-*  @return      The read byte
-*/
-extern "C" unsigned char inb(unsigned short port);
-
-#endif /* INCLUDE_IO_H */
